Add editor command utils
Make the following changes to make creating a menu for the editor
easier:

Editor now takes children, you'd use it as follows:
```tsx
function App() {
    return (
      <Editor>
        <Menu />
      </Editor>
    );
}
```

Inside the Editor component, you can use the functions `useCommand`,
which returns whether a command is available with the current editor
state, or `dispatchCommand`, which will dispatch a command using the
current editor state.

You can also use `useEditorState` to get the current state of the
editor.