Extend parser to support full MyST suite
This commit extends the parser, and uses it like done in the MyST demo.
Unfortunately, this requires quite a bit of plugin configuration, and
future extensions should be put in here as well.

Tests need to be fixed up as well, as the second stage of AST processing
puts everything in a top-level "block" element.