Parser: add tests and fix missing nodes
I added more tests now, to also do a round-trip.

Should add some extra documents to test on as well.

Unfortunately, round-trip does not convert back things to the way they
were exactly. This is because there are multiple ways to represent
formatting in Myst, like ``` and ::: for directives, or math`E=mc^2` and
$E=mc^2$ for math equations. The workaround for this would be a diffing
algorithm that compares the AST of the original source with the new one,
and only updates nodes that have changed. However, this is convoluted,
because we'd have to merge text nodes as well.